+++ public/stylesheets/style.css
@@ -3,21 +3,38 @@ body {
   padding-bottom: 40px;
 }
 
+button:focus {
+    outline: 0 !important;
+}
+
 #goMap {
   margin-top: 10px;
   padding: 10px;
 }
 
+#locationName {
+  width: 300px;
+}
+
+#map {
+  width: 100%;
+  height: 400px;
+}
+
 .first-jumbo {
-  background-image:url('/images/tropical-background.jpg');
-  background-size: cover; 
+  background-image: url('/images/tropical-background.jpg');
+  background-size: cover;
 }
 
 .second-jumbo {
-background: #DD5E89;  /* fallback for old browsers */
-background: -webkit-linear-gradient(to left, #F7BB97, #DD5E89);  /* Chrome 10-25, Safari 5.1-6 */
-background: linear-gradient(to left, #F7BB97, #DD5E89); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
+  background: #DD5E89;
+  /* fallback for old browsers */
+  background: -webkit-linear-gradient(to left, #F7BB97, #DD5E89);
+  /* Chrome 10-25, Safari 5.1-6 */
+  background: linear-gradient(to left, #F7BB97, #DD5E89);
+  /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
 }
+
 .glyphicon-star {
   color: #DAA520;
 }
@@ -25,20 +42,27 @@ background: linear-gradient(to left, #F7BB97, #DD5E89); /* W3C, IE 10+/ Edge, Fi
 .messages {
   align-content: left !important;
 }
+
 .navbar-brand {
   text-transform: uppercase;
   letter-spacing: 5px;
 }
 
+.navbar {
+  margin-bottom: 10px;
+  border-radius: 0px;
+}
+
 .front-page-text {
-  color: white !important; 
+  color: white !important;
 }
 
 .img-responsive {
   max-height: 350px;
 }
 
-#currency-selector, #language-selector {
+#currency-selector,
+#language-selector {
   width: 100px;
   text-align: center;
 }
@@ -56,67 +80,33 @@ background: linear-gradient(to left, #F7BB97, #DD5E89); /* W3C, IE 10+/ Edge, Fi
   padding: 0;
 }
 
-/* FORMS */
 
+/* SETTINGS FOR ALL FORMS */
 .form-signin {
   max-width: 400px;
   padding: 15px;
   margin: 0 auto;
 }
+
 .form-signin .form-signin-heading,
 .form-signin .checkbox {
   margin-bottom: 10px;
 }
+
 .form-signin .checkbox {
   font-weight: normal;
 }
+
 .form-signin .form-control {
   position: relative;
   height: auto;
   -webkit-box-sizing: border-box;
-     -moz-box-sizing: border-box;
-          box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
   padding: 10px;
   font-size: 16px;
 }
+
 .form-signin .form-control:focus {
   z-index: 2;
-}
-
-.navbar {
-  margin-bottom: 10px;
-  border-radius: 0px;
-}
-
-#locationName {
-  width: 300px;
-}
-
-#map {
-    width: 100%;
-    height: 400px;
-}
-
-.controls {
-    margin-top: 10px;
-    border: 1px solid transparent;
-    border-radius: 2px 0 0 2px;
-    box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    height: 32px;
-    outline: none;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3);
-}
-#searchInput {
-    background-color: #fff;
-    font-family: Roboto;
-    font-size: 15px;
-    font-weight: 300;
-    margin-left: 12px;
-    padding: 0 11px 0 13px;
-    text-overflow: ellipsis;
-    width: 50%;
-}
-#searchInput:focus {
-    border-color: #4d90fe;
 }
