+++ routes/auth.js
@@ -4,8 +4,6 @@
 
 const express = require('express');
 const bcrypt = require('bcrypt');
-// const passport = require('passport');
-
 const passport = require('passport');
 
 const router = express.Router();
