+++ routes/chat.js
@@ -1,16 +1,9 @@
-// const passport = require('passport');
-
-// const AuthenticationController = require('../controllers/authentication');
-// const UserController = require('../controllers/user');
-// const ChatController = require('../controllers/chat');
 const express = require('express');
 
 const auth = require('../helpers/auth.js');
 const Conversation = require('../models/conversation');
 const Message = require('../models/message');
 const Experience = require('../models/experience');
-// const Chat = require('../models/chat');
-// const User = require('../models/user');
 
 const router = express.Router();
 
@@ -20,30 +13,24 @@ router.post('/:idexp/new/:recipient', (req, res, next) => {
     res.status(422).send({ error: 'Please choose a valid recipient for your message.' });
     return next();
   }
-
   if (!req.body.composedMessage) {
     res.status(422).send({ error: 'Please enter a message.' });
     return next();
   }
-  // Conversation.find({ experience: req.params.idexp, participants: req.params.recipient, participants: req.user._id }
-
   const conversation = new Conversation({
     participants: [req.user._id, req.params.recipient],
     experience: req.params.idexp,
   });
-
   conversation.save((err, newConversation) => {
     if (err) {
       res.send({ error: err });
       return next(err);
     }
-
     const message = new Message({
       conversationId: newConversation._id,
       body: req.body.composedMessage,
       author: req.user._id,
     });
-
     message.save((err, newMessage) => {
       if (err) {
         res.send({ error: err });
@@ -56,38 +43,29 @@ router.post('/:idexp/new/:recipient', (req, res, next) => {
         if (err) {
           next(err);
         } else {
-          // console.log(result);
           res.render('chats/show', { result, newConversation, newMessage });
         }
       });
-      // res.status(200).json({ message: 'Conversation started!', conversationId: conversation._id });
-      // return next();
     });
   });
 });
 
 // Send reply in conversation
 router.post('/:conversationId', (req, res, next) => {
-  // console.log('req.body', req.body);
-  // console.log('req.params', req.params);
   const reply = new Message({
     conversationId: req.params.conversationId,
     body: req.body.composedMessage,
     author: req.body.idsender,
   });
-  // console.log('reply', reply);
-
   reply.save((err, sentReply) => {
     if (err) {
       res.send({ error: err });
       return next(err);
     }
-
     res.status(200).json({ message: 'Reply successfully sent!' });
     return (next);
   });
 });
-
 
 // View messages to and from authenticated user
 router.get('/', (req, res, next) => {
@@ -115,7 +93,6 @@ router.get('/', (req, res, next) => {
             }
             fullConversations.push(message);
             if (fullConversations.length === conversations.length) {
-
               res.render('chats/index', { conversations: fullConversations });
             }
           });
@@ -148,14 +125,13 @@ router.get('/:expId/:conversationId', (req, res, next) => {
         if (err) {
           next(err);
         } else {
-      // res.status(200).json({ conversation: messages });
           res.render('chats/show', { conversation: messages, result, newMessage, messages });
         }
       });
     });
 });
 
-// Retrieve single conversation
+// Retrieve single conversation - API JSON
 router.get('/:conversationId', (req, res, next) => {
   const expId = req.params.expId;
   const conversationId = req.params.conversationId;
@@ -177,6 +153,5 @@ router.get('/:conversationId', (req, res, next) => {
       res.status(200).json({ conversation: messages });
     });
 });
-
 
 module.exports = router;
