--- socketEvents.js
@@ -1,25 +1,0 @@
-exports = module.exports = (io) => {
-  // Set socket.io listeners.
-  io.on('connection', (socket) => {
-    console.log('a user connected');
-
-    // On conversation entry, join broadcast channel
-
-    socket.on('enter conversation', (conversation) => {
-      socket.join(conversation);
-      console.log(`joined ${conversation}`);
-    });
-
-    socket.on('leave conversation', (conversation) => {
-      socket.leave(conversation);
-      console.log(`left ${conversation}`);
-    });
-    socket.on('chat message', (conversation) => {
-      io.sockets.in(conversation).emit('refresh messages', conversation);
-    });
-
-    socket.on('disconnect', () => {
-      // console.log('user disconnected');
-    });
-  });
-};
