+++ views/auth/signup.ejs
@@ -1,20 +1,20 @@
 <form class="form-signin" action="/signup" method="POST">
   <h2 class="form-signin-heading">Please sign up</h2>
   <a href="/auth/facebook">Sign up with help of your Facebook account</a><br><br>
-<div class="input-group margin-bottom-sm">
-  <span class="input-group-addon"><i class="fa fa-envelope-o fa-fw"></i></span>
-  <input class="form-control" type="email" name="email" placeholder="Email address" required autofocus>
+  <div class="input-group margin-bottom-sm">
+    <span class="input-group-addon"><i class="fa fa-envelope-o fa-fw"></i></span>
+    <input class="form-control" type="email" name="email" placeholder="Email address" required autofocus>
   </div>
   <div class="input-group">
     <span class="input-group-addon"><i class="fa fa-key fa-fw"></i></span>
     <input class="form-control" name="password" type="password" placeholder="Password" required>
   </div>
-    <div class="input-group">
+  <div class="input-group">
     <span class="input-group-addon"><i class="fa fa-user fa-fw"></i></span>
     <input class="form-control" name="name" type="text" placeholder="Name" required>
   </div><br>
-      <br>
-        <% if (typeof(errorMessage) != "undefined") { %>
+  <br>
+  <% if (typeof(errorMessage) != "undefined") { %>
     <div class="error-message">
       <%= errorMessage %>
     </div>
