+++ views/experiences/index.ejs
@@ -2,25 +2,26 @@
   <h1>Experiences</h1>
   <hr>
 
-    <div class="row">
-      <% result.forEach(experience => { %>
-        <a href="/experiences/<%= experience._id %>/"><div class="col-sm-6 col-md-4">
+  <div class="row">
+    <% result.forEach(experience => { %>
+      <a href="/experiences/<%= experience._id %>/">
+        <div class="col-sm-6 col-md-4">
           <div class="thumbnail" style="min-height: 400px !important;">
             <img class="front-image-thumbnail" src="<%= experience.images[0] %>" alt="<%= experience.name %>">
             <div class="caption">
               <h3>
-                
-                  <%= experience.name %>
+
+                <%= experience.name %>
               </h3><br></a>
-              <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-              <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-              <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-              <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-              <span class="glyphicon glyphicon-star" aria-hidden="true"></span><br>
-              <p><strong>$<%= experience.price %></strong></p>
-            </div>
-          </div>
-        </div>
-        <% }); %>
-    </div>
+      <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+      <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+      <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+      <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+      <span class="glyphicon glyphicon-star" aria-hidden="true"></span><br>
+      <p><strong>$<%= experience.price %></strong></p>
+      </div>
+      </div>
+      </div>
+      <% }); %>
   </div>
+</div>
