+++ views/index.ejs
@@ -65,11 +65,21 @@
         <li>
           <h2>Must-See Attractions</h2>
         </li>
-        <li><h4><a href="#">11-hour guided day trip from London to Windsor Castle, Stonehenge and Bath</a></h4></li>
-        <li><h4><a href="#">Explore Gaudí and modernisme</a></h4></li>
-        <li><h4><a href="#">Eiffel Tower Summit and VIP Lounge Priority Access</a></h4></li>
-        <li><h4><a href="#">Carnival in Rio de Janeiro</a></h4></li>
-        <li><h4><a href="#">The Vasa Museum</a></h4></li>
+        <li>
+          <h4><a href="#">11-hour guided day trip from London to Windsor Castle, Stonehenge and Bath</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Explore Gaudí and modernisme</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Eiffel Tower Summit and VIP Lounge Priority Access</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Carnival in Rio de Janeiro</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">The Vasa Museum</a></h4>
+        </li>
       </ul>
     </div>
     <div class="col-sm-6 col-md-4">
@@ -77,11 +87,21 @@
         <li>
           <h2>Top Destinations</h2>
         </li>
-        <li><h4><a href="#">Barcelona, Spain</a></h4></li>
-        <li><h4><a href="#">Paris, France</a></h4></li>
-        <li><h4><a href="#">Rome, Italy</a></h4></li>
-        <li><h4><a href="#">Rio de Janeiro, Brazil</a></h4></li>
-        <li><h4><a href="#">Stockholm, Sweden</a></h4></li>
+        <li>
+          <h4><a href="#">Barcelona, Spain</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Paris, France</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Rome, Italy</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Rio de Janeiro, Brazil</a></h4>
+        </li>
+        <li>
+          <h4><a href="#">Stockholm, Sweden</a></h4>
+        </li>
       </ul>
     </div>
     <div class="col-sm-6 col-md-4">
@@ -114,4 +134,5 @@
 <!-- /container -->
 <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAzZ_vMmKMZeivRsR3pp-ZM4cWviVEkG1c&libraries=places" async
   defer></script>
-<script src="http://localhost:3000/javascripts/searchexperience.js"></script>
+<script src="/javascripts/init-getCoordinates-textSearch.js"></script>
+<script src="/javascripts/searchexperience.js"></script>
