+++ views/profile/show.ejs
@@ -1,37 +1,48 @@
 <div class="container">
 
-<h1><%= user.name %></h1>
-<hr>
-<img class="img-circle img-responsive profile-img" src="<%= user.pic_path %>"><br>
-<% if (currentUser.location.city !== undefined ) { %>
-  <h3><%= user.location.city%>, <%= user.location.country %></h3>
-<% } %> 
+  <h1>
+    <%= user.name %>
+  </h1>
+  <hr>
+  <img class="img-circle img-responsive profile-img" src="<%= user.pic_path %>"><br>
+  <% if (currentUser.location.city !== undefined ) { %>
+    <h3>
+      <%= user.location.city%>,
+        <%= user.location.country %>
+    </h3>
+    <% } %>
+      <p>
+        <%= user.about %>
+      </p>
+      <a class="btn btn-warning" href="/profile/<%=user._id%>/edit">Edit Profile</a>
+      <a href="/experiences/new" class="btn btn-success">Add Experience</a><br>
+
+      <hr>
+      <h3>My Experiences</h3>
+      <% user.experiences.forEach(experience => { %>
+        <h4>
+          <a href="/experiences/<%= experience._id %>">
+            <%= experience.name %>
+          </a>
+        </h4>
+        <strong>Description: </strong>
+        <%= experience.description %><br>
+          <img class="img-responsive" src="<%= experience.images[0] %>"><br><br>
+          <a class="btn btn-warning" href="/experiences/<%= experience._id %>/edit">Edit</a>
+          <a class="btn btn-danger" href="/experiences/<%= experience._id %>/delete">Delete</a>
+          <hr>
+          <% }) %>
+            <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+            <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+            <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+            <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+            <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
+            <h3>Reviews From Experifiers (63)</h3><br>
 
 
-<p><%= user.about %></p>
-<a class="btn btn-warning" href="/profile/<%=user._id%>/edit">Edit Profile</a> 
-<a href="/experiences/new" class="btn btn-success">Add Experience</a><br>
-
-<hr>
-<h3>My Experiences</h3>
-<% user.experiences.forEach(experience => { %>
-<h4><a href="/experiences/<%= experience._id %>"><%= experience.name %></a></h4>
-<strong>Description: </strong><%= experience.description %><br>
-<img class="img-responsive" src="<%= experience.images[0] %>"><br><br>
-<a class="btn btn-warning" href="/experiences/<%= experience._id %>/edit">Edit</a> 
-<a class="btn btn-danger" href="/experiences/<%= experience._id %>/delete">Delete</a>
-<hr>
-<% }) %>
-<span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-          <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-          <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-          <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-          <span class="glyphicon glyphicon-star" aria-hidden="true"></span>
-<h3>Reviews From Experifiers (63)</h3><br>
-
-
-<a href="#"><img style="zoom: 50%;" src="http://stephanvandijk.com/img/profile_round.png"><br><strong>John</strong></a>
-<p>Good intimate jazz performance. Started on time but the band, singer and fellow travelers were not introduced as one would think for this type of experience. The house dog roaming and trying to eat guests food was also distracting.</p>
-...
-... 
+            <a href="#"><img style="zoom: 50%;" src="http://stephanvandijk.com/img/profile_round.png"><br><strong>John</strong></a>
+            <p>Good intimate jazz performance. Started on time but the band, singer and fellow travelers were not introduced
+              as one would think for this type of experience. The house dog roaming and trying to eat guests food was also
+              distracting.</p>
+            ... ...
 </div>
