--- views/reviews/edit.ejs
@@ -1,16 +1,0 @@
-<div class="container">
-    <div class="row">
-        <h1 style="text-align: center;">Edit Review</h1>
-        <div style="width: 30%; margin: 30px auto;">
-        <form action="/experiences/<%= experience_id %>/reviews/<%= review._id %>?_method=PUT" method="POST">
-            <div class="form-group">
-            <input class="form-control" type="text" name="review[text]" value="<%= review.text %>">
-            </div>
-            <div class="form-group">
-            <button class="btn btn-lg btn-primary btn-block">Submit!</button>
-            </div>
-        </form>
-                <a href="/experiences">Go Back</a>
-        </div>
-    </div>
-</div>
