--- views/reviews/new.ejs
@@ -1,16 +1,0 @@
-<div class="container">
-    <div class="row">
-        <h1 style="text-align: center;">Add new Review to <%= experience.name %></h1>
-        <div style="width: 30%; margin: 30px auto;">
-        <form action="/experiences/<%= experience._id %>/reviews" method="POST">
-            <div class="form-group">
-            <input class="form-control" type="text" name="review[text]" placeholder="text">
-            </div>
-            <div class="form-group">
-            <button class="btn btn-lg btn-primary btn-block">Submit!</button>
-            </div>
-        </form>
-                <a href="/reviews">Go Back</a>
-        </div>
-    </div>
-</div>
